backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.7.2
scikit-learn==1.3.2
python-dotenv==1.0.0
pydantic==2.5.2
numpy==1.26.2
httpx==0.25.2
pypdf>=4.0.0
python-multipart>=0.0.6
